Finish changing types

+++ src/rw_json_parser.h
@@ -16,14 +16,14 @@
 ** \param idx The index of the character '[' that begins the current rw_array_t
 ** \returns The json rw_array_t parsed from the position
 */
-rw_array_t *rw_parse_array_buff(char *b, unsigned long *idx);
+rw_array_t *rw_parse_array_buff(char *b, size_t *idx);
 /**
 ** \param f The file stream
 ** \param pos The postion in the file of the character after the '[' that
 **            begins the current rw_array_t
 ** \returns The json rw_array_t parsed from the position
 */
-rw_array_t *rw_parse_array(FILE *f, unsigned long *pos);
+rw_array_t *rw_parse_array(FILE *f, size_t *pos);
 
 /**
 ** \param b The buffer containing the object currently being parsed
@@ -35,7 +35,7 @@ rw_array_t *rw_parse_array(FILE *f, unsigned long *pos);
 **            index starts at 0
 ** \returns The json dict parsed from the index
 */
-rw_dict_t *rw_parse_dict_buff(char *b, unsigned long *idx);
+rw_dict_t *rw_parse_dict_buff(char *b, size_t *idx);
 /**
 ** \param f The file stream
 ** \param pos A pointer to the position in the file of the character after
@@ -43,7 +43,7 @@ rw_dict_t *rw_parse_dict_buff(char *b, unsigned long *idx);
 **            '{' that begins the current dict
 ** \returns The json dict parsed from the position
 */
-rw_dict_t *rw_parse_dict(FILE *f, unsigned long *pos);
+rw_dict_t *rw_parse_dict(FILE *f, size_t *pos);
 
 /**
 ** \brief Parse the given file and returns the associated JSON object
